hf_spaces/rag/requirements.txt
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
huggingface-hub>=0.24.0
transformers>=4.43.0
torch>=2.2.0
accelerate>=0.30.0
numpy>=1.24.0
